add .gitignore and first script
I created this repository to give Ryan some sample code to
reference. The file `runpass.py' uses the cfbd API to get data
on Notre Dame's plays from 2023 and calculate the number of runs
and passes after a run or pass play. It's not pretty but the idea
is to create code that works and is simple.